feat(bixarena): restructure global styles into lib partials
