+++ apps/bixarena/app-angular/src/styles.scss
@@ -1,47 +1,1 @@
-@use 'bixarena/styles/src/lib/variables';
-@use '@fontsource-variable/dm-sans';
-
-// Layer order: reset < primeng < app (unscoped rules beat all layers)
-@layer reset, primeng, app;
-
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  background: var(--p-surface-0);
-  color: var(--p-text-color);
-  font-family: 'DM Sans Variable', 'DM Sans', system-ui, sans-serif;
-  font-size: 0.9375rem;
-  line-height: 1.6;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-// PrimeNG components read --p-font-family; body font-family alone doesn't propagate into them
-:root {
-  --p-font-family: 'DM Sans Variable', 'DM Sans', system-ui, sans-serif;
-}
-
-.container {
-  width: 100%;
-  max-width: var(--max-width);
-  margin-inline: auto;
-  padding-inline: 2rem;
-}
-
-@layer app {
-  bixarena-nav .p-button {
-    padding: 0 0.875rem;
-    font-size: 0.875rem;
-    font-family: inherit;
-    font-variation-settings: 'wght' 500;
-    letter-spacing: -0.01em;
-    border-radius: 6px;
-    height: 32px;
-    line-height: 32px;
-  }
-}
+@use 'bixarena/styles/src/lib/index';
