+++ libs/bixarena/styles/src/lib/_overrides.scss
@@ -1,0 +1,13 @@
+// PrimeNG component style overrides.
+@layer app {
+  bixarena-nav .p-button {
+    padding: 0 0.875rem;
+    font-size: 0.75rem;
+    font-family: inherit;
+    font-variation-settings: 'wght' 540;
+    letter-spacing: -0.01em;
+    border-radius: var(--p-border-radius-md);
+    height: 1.875rem;
+    line-height: 1.875rem;
+  }
+}
