+++ libs/bixarena/styles/src/lib/_utilities.scss
@@ -1,0 +1,7 @@
+// Reusable layout helper classes.
+.container {
+  width: 100%;
+  max-width: var(--max-width);
+  margin-inline: auto;
+  padding-inline: 2rem;
+}
